fix: delay the device/connect init, after the analytics, so THP autoconnect is not triggered too soon
